Help! pls add ShopScene to "3" in Build Setting! Thanks!
Add ShopScene to "3" in Build Setting pls! I have no permission to change Build Setting, but I have to use Build Setting to make scene work. Thank you.
Add the button to jump to shop. but the choice of code function in unity not show well in my side. seems like unity 2021.3.18f1 don't have this function. Does anyone want to use unity 2022 to try? It might break the file but it have function to add action on button.